+++ lib/src/components/ec.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/widgets.dart';
 import 'package:ultimate_containers/ultimate_containers.dart';
 
-/// Expanded Ultimate Container
 class EC extends StatelessWidget {
+  /// Expanded Ultimate Container
   const EC(
     this.inputs, {
     Key? key,